requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.5
django-tinymce==3.5.0
django-user-agents==0.4.0
dnspython==2.3.0
email-validate==1.1.2
executing==1.2.0
filelock==3.9.0
idna==2.10
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.1
transliterate==1.10.2
tzdata==2022.7
ua-parser==0.16.1
user-agents==2.2.0
wcwidth==0.2.6
